requirements.txt
Flask==3.0.0
pymongo==4.6.1
requests
flask-sqlalchemy
gunicorn
pymysql